# simpler perf reduce string