--- a/include/ada/checkers.h
+++ b/include/ada/checkers.h
@@ -54,6 +54,8 @@ namespace ada::checkers {
     return view.size() >= prefix.size() && (view.substr(0, prefix.size()) == prefix);
   }
 
+  ada_really_inline constexpr bool is_ipv4(std::string_view view) noexcept;
+
 } // namespace ada::checkers
 
 #endif //ADA_CHECKERS_H