--- a/src/ada.cpp
+++ b/src/ada.cpp
@@ -1,4 +1,5 @@
 #include "ada.h"
+#include "checkers.cpp"
 #include "unicode.cpp"
 #include "serializers.cpp"
 #include "implementation.cpp"