This is an alternative to the perf reduce string branch.

The gist of the change is that instead of doing...

```
url.is_valid = parse_host(url.host, host_view, !url.is_special(), is_ascii);
```

We just do...

```C++
url.parse_host(host_view)
```

You can get much the same performance benefits, with more maintainable code.

I did not code it for performance benefits, but there are some...

```
Main branch (updated just now)
BasicBench_AdaURL                     5442 ns         5442 ns       128532 time/byte=6.73507ns time/url=494.721ns url/s=2.02134M/s

This PR:
BasicBench_AdaURL                     4662 ns         4662 ns       150047 time/byte=5.76945ns time/url=423.792ns url/s=2.35965M/s
```


Note: this is a PR against main, but built on top of @anonrig 's PR. Yet because we have done more work in-between, it is no longer a clean PR-on-top-of-PR, so a propose to go to main directly.